requirements.txt
# Core LLM + agent framework (pinned for stability)
langchain==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-community==0.2.16

# UI
gradio

# Environment & validation
python-dotenv
pydantic
